Allow for possibility of Bundle response in payer suite
The payer suite assumed the response to the  questionnaire-package
operation was a Parameters instance, but it can instead be a Bundle.
This change accounts for both possibilities.

There are a lot of changes, because there are many places in the
code that relied on the assumption that the request was a Parameters.